Improve docker image and docs

+++ agilepy/scripts/utilities.sh
@@ -1,3 +1,8 @@
+: '
+    $1:$2 is the repository:tag of the image to be boostrapped.
+    $3 is DUPLICATE_IMAGE. If false, the original image will be overwritten.
+'
+
 boostrap() {
     MY_UID="$(id -u)"
     MY_GID="$(id -g)"
@@ -9,9 +14,20 @@ boostrap() {
         printf "\n\33[31mImage $AGILEPY_TAG does not exist locally, pull it from DockerHub.\n\33[0m\n"
         return;
     fi
+    
+    docker tag $AGILEPY_TAG $INTERMEDIATE_TAG
 
-docker tag $AGILEPY_TAG $INTERMEDIATE_TAG
-docker rmi $AGILEPY_TAG
+    # check if $3 is false
+    if [[ $3 == false ]]; then
+        docker rmi $AGILEPY_TAG
+        printf "\n\33[31mThe original image will be deleted.\33[0m\n"
+    else
+        AGILEPY_TAG=$AGILEPY_TAG\_$(whoami)
+        printf "\n\33[32mThe original image will be kept. The new image will have the following tag: '$AGILEPY_TAG'\33[0m\n"        
+    fi
+    
+    printf "\n\33[32mBuilding image $AGILEPY_TAG\33[0m\n"
+    
 docker build \
 --tag $AGILEPY_TAG \
 - <<EOF
